+++ gd-smartfarm/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## UNRELEASED
+
+### Added
+- Added header-footer layout and empty components for each page. [SMARTFARM-8](https://opensource.ncsa.illinois.edu/jira/browse/SMARTFARM-8)
+
 ## [0.1.0] - 2020-10-01
 ### Added
-- Implemented the basic layout.
+- Implemented the basic setup.